Added write to config file. Finished customisation options. Updated UI.
Everything is now updated to the config file. So users settings and customisations stay when app is relaunched. Customisation options are finished, users can change the theme, scale, button colours and button hover colours. This is all saved with error handling. Note: if any issues ever arise, deleting the config.txt file will likely fix it. Added a todo for error checking config file loading. Shouldn't be an issue unless config.txt is corrupted or the user modifies it. The program should write to it correctly.